FIX: 코스 조회 응답의 고도값 수정 (#88)
코스 전체,상세 조회 API 응답의 고도값을 자연수 형태로 반환하도록 수정했습니다.
관련 테스트 코드도 함께 수정했습니다.

+++ src/main/java/com/server/running_handai/domain/course/dto/CourseInfoWithDetailsDto.java
@@ -1,6 +1,5 @@
 package com.server.running_handai.domain.course.dto;
 
-import com.server.running_handai.domain.course.entity.Course;
 import java.util.List;
 
 public record CourseInfoWithDetailsDto(
@@ -9,7 +8,7 @@ public record CourseInfoWithDetailsDto(
         String thumbnailUrl,
         double distance,
         int duration,
-        double maxElevation,
+        int maxElevation,
         double distanceFromUser,
         int bookmarks,
         boolean isBookmarked,
@@ -22,7 +21,7 @@ public record CourseInfoWithDetailsDto(
                 courseInfoDto.getThumbnailUrl(),
                 courseInfoDto.getDistance(),
                 courseInfoDto.getDuration(),
-                courseInfoDto.getMaxElevation(),
+                (int) courseInfoDto.getMaxElevation(),
                 courseInfoDto.getDistanceFromUser(),
                 bookmarks,
                 isBookmarked,
